[MOV] spreadsheet(_dashboard): rename `filter_search_dialog`
This commit renames the `filter_search_dialog` files to
`filter_values_component`. "Search dialog" doesn't make sense, as it
is not a dialog anymore, and we never searched anything in it.

Task: 4816189
