+++ addons/spreadsheet/static/src/global_filters/components/filter_values_list/filter_values_list.scss
